Move event background updates to a separate file
